fix: suppress -Wconversion/-Wsign-conversion/-Wdouble-promotion in ww3d2
Add explicit static_cast<> at all type-narrowing and signedness-change
sites in Code/ww3d2/ and incidental fixes in Scripts/:

- HRESULT (long) → uint32_t: DX8_ErrorCode() call sites in
  assetmgr, ddsfile, dx8indexbuffer, dx8vertexbuffer, dx8wrapper,
  missingtexture, surfaceclass, texture, textureloader, ww3d.cpp
- unsigned long / long → int/uint: bwrender (CRC, indices),
  dx8renderer (CRC), meshmatdesc (CRC), render2d, shader (_PolygonCullMode),
  stripoptimizer (ptrdiff_t), visrasterizer (Float_To_Long),
  w3d_dep (ptrdiff_t), ww3d (RECT fields, swap interval)
- size_t → int/uint: animatedsoundmgr, ddsfile, hmorphanim,
  meshgeometry, meshmdlio, rendobj, texturethumbnail, vertmaterial
- float → int: colorspace, hmorphanim, mesh, motchan, part_emt,
  render2dsentence, ringobj, sphereobj, visrasterizer
- float → uint8_t: dx8wrapper.h, font3d
- float → bool: dazzle, meshmatdesc, part_buf, part_emt, rendobj
- float → uint32_t: motchan, part_buf
- double → float: camera, hanim, htree, intersec.inl, light,
  light.h, lightenvironment, matrixmapper, meshbuild, shattersystem,
  txt (sqrt→sqrtf, ceil→ceilf)
- Fix \\!= corruption (zsh history expansion via heredoc) in
  meshmatdesc, Scripts/Mission02, Scripts/Toolkit_Powerup

+++ original/Code/ww3d2/animatedsoundmgr.cpp
@@ -63,7 +63,7 @@ Build_List_From_String
 		 (delimiter != NULL) &&
 		 (string_list != NULL))
 	{
-		int delim_len = ::strlen (delimiter);
+		int32_t delim_len = static_cast<int32_t>(::strlen (delimiter));
 
 		//
 		// Determine how many entries there will be in the list
