+++ original/Code/ww3d2/meshbuild.cpp
@@ -1075,14 +1075,14 @@ void MeshBuilderClass::Compute_Bounding_Sphere(Vector3 * set_center,float * set_
 			radsqr = radius * radius;
 
 			double oldtonew = testrad - radius;
-			center.X = (radius * center.X + oldtonew * Verts[i].Position.X) / testrad;
-			center.Y = (radius * center.Y + oldtonew * Verts[i].Position.Y) / testrad;
-			center.Z = (radius * center.Z + oldtonew * Verts[i].Position.Z) / testrad;
+			center.X = static_cast<float>((radius * center.X + oldtonew * Verts[i].Position.X) / testrad);
+			center.Y = static_cast<float>((radius * center.Y + oldtonew * Verts[i].Position.Y) / testrad);
+			center.Z = static_cast<float>((radius * center.Z + oldtonew * Verts[i].Position.Z) / testrad);
 		}
 	}
 
 	*set_center = center;
-	*set_radius = radius;
+	*set_radius = static_cast<float>(radius);
 }
 
 /***********************************************************************************************
