+++ original/Code/ww3d2/ringobj.cpp
@@ -294,7 +294,7 @@ void RingRenderObjClass::Generate_Shared_Mesh_Arrays (void)
 
 		for(int i=0; i < RING_NUM_LOD; i++) {
 
-			RingMeshArray[i].Generate(1.0f, size);
+			RingMeshArray[i].Generate(1.0f, static_cast<int32_t>(size));
 
 			size+=step;
 		}
@@ -585,7 +585,7 @@ void RingRenderObjClass::Render(RenderInfoClass & rinfo)
 		screen_size = sqrtf(screen_size);
 
 		float lod     = screen_size * ((float) RING_NUM_LOD);
-		int	lod_int = lod;
+		int32_t	lod_int = static_cast<int32_t>(lod);
 		lod-=lod_int;
 
 		if (lod >= 0.5f) lod_int++;
